Fix float to decimal problem on AddProductSaleView model.

+++ SUD/ViewModels/AddProductSaleView.cs
@@ -20,11 +20,11 @@ namespace SUD.ViewModels
         [Required(ErrorMessage = "Cantidad Requerida")]
         [Display(Name = "Porcentaje de IVA", Description = "0.00")]
         [RegularExpression(@"^[0-9]\d*(\.\d +)?$", ErrorMessage = "Porcentaje Invalido, solo se permiten numeros enteros.")]
-        public float IVAPercentage { get; set; }
+        public decimal IVAPercentage { get; set; }
 
         [Required(ErrorMessage = "Cantidad Requerida")]
         [Display(Name = "Porcentaje de descuento", Description = "0.00")]
         [RegularExpression(@"^[0-9]\d*(\.\d +)?$", ErrorMessage = "Porcentaje Invalido, solo se permiten numeros enteros.")]
-        public float DiscountRate { get; set; }
+        public decimal DiscountRate { get; set; }
     }
 }
